backend/requirements.txt

fastapi
uvicorn[standard]
python-multipart
tensorflow
Pillow
numpy
python-jose[cryptography]
passlib[bcrypt]
